requirements.txt
fastapi
uvicorn
python-multipart
motor>=3.7.1
httpx
pydantic-settings
python-dotenv
google-genai
sentence-transformers>=2.2.2
faiss-cpu==1.13.2
pymongo>=4.10.0
requests>=2.32.5,<3.0.0
PyPDF2>=3.0.1
openpyxl==3.1.2
pandas>=2.2.3
numpy>=2.1.0
pyjwt
